Benchmark Feature Normalization (#1614)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1614

Machine learning is becoming an important use case for Velox, and a lot of work is
going into optimizing Velox for it.

Adding a benchmark for normalizing float features which is a fairly standard operation
and one which we want to make sure we don't regress on.

Results as of today on my machine

============================================================================
[...]hmarks/basic/FeatureNormalization.cpp     relative  time/iter   iters/s
============================================================================
normalize                                                  26.69ns    37.46M
normalizeConstant                                          33.05ns    30.26M

Differential Revision: D36382333

fbshipit-source-id: fe492a79ebe77ee443e662d950c51bd26a04a6a3