Refactor SeekableInputStream::loadIndices()
loadIndices() in SeekableInputStream returns the the positions array
index for the next stream in the current column for DWRF, relative to
the startIndex parameter. It was used in
SelectiveStringDictionaryColumnReader to lazily load the stride
dictionary for DWRF. To prepare for the upcoming refactoring that
moves this class and its subclasses to dwio::common, we need to remove
the DWRF related concept from it. This commit changed loadIndices() to
skipPositions() which takes a PositionProvider parameter, and would
skip the number of positions needed by this stream to identify any
stream address.