Async Vector Function (#1665)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1665

Those are initial draft classes that illustrates the way we would implement the Async support
for remote UDFs in Velox.

This diff adds the AsynVectorFunction which has the following functions:
1. asyncApply(): provided by the user (or the AsyncFunctionAdapter).
2. apply(): not provided by the user, calls asyncApply and blocks wait for asyncApply().
3. threadSafe(): can we schedule multiple vector functions to run on different threads concurrently.
this would be inspected by the async expression evaluator if it does multi-threaded scheduling.

This function

The next steps are:
1. prototype for AsyncFunctionAdapter: (convert simple function to AsynVectorFunction).
2. prototype for async expression eval: evaluate a tree of AsynVectorFunction.
3. proto-type a pull model with async internals (koski-request).

4. try to add it to cmake and compile it in clang oss. from coro documentation:
```
Currently (c. mid 2020) the only compiler with a production-quality implementation of coroutines compatible with folly::coro is Clang.
```

Differential Revision: D36549995

fbshipit-source-id: 14b2ea6a025b352dce47e92ff97999a6e43e9f7a