Refactor compiler build flags to one location (#1667)
Summary:
Currently Velox build flags are sprinkled across not only our code base , but several dependent code bases. Since Velox is targeting multiple cpu architectures, it makes sense to have one single source to determine the appropriate flags to use through out this and dependent projects.
This pr moves this logic to a shell function  so we can get the right flags for the setup scripts, and refactors scripts to call this.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1667

Reviewed By: kagamiori

Differential Revision: D36556332

Pulled By: kgpai

fbshipit-source-id: 3ae8b93e5bb4b992185444ab381dbff9542e2b3a