Remove extra calls to SelectivityVector::updateBounds()
Do not call updateBounds when the bounds are either known or will be
overwritten. Add a resizeFill method that sets the size and contents
and does not call updateBounds.

Add LocalSelectivityVector interfaces that initialize the vector from
another vector or from size and value.

Fix SelectivityVector::== to only consider bits in between begin_ and
end_.