Add Task::maybeReserve (#1681)
Summary:
Add MemoryUsageTracker::maybeReserve

Adds a method that tries to increase a reservation in a
MemoryUsageTracker if there seems to be space. If so, this increases
the reservation and catches an eventual error. Returns true if the
reservation was increased.

This will be extended when adding Task memory cap negotiation.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1681

Reviewed By: mbasmanova

Differential Revision: D36611909

Pulled By: oerling

fbshipit-source-id: 693b85b7398b3ffce4f84aa95b55c40314b4faff