Fix LocalPartition to return proper output type in output vectors. (#1698)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1698

"Field not found" error has been triggered for certain queries with "union all".
In such queries LacalPartition was retuning output vectors (data for the next operator) with not the output type, but the type which was set by the source (to LocalPartiton) operator.
This change makes sure we use proper output type in the returned output vector.

Differential Revision: D36654248

fbshipit-source-id: af9ffb8b5c38640df02c1ea56c720e976169d3fe