Remove extra init from SelectivityVector (#1672)
Summary:
Remove extra calls to SelectivityVector::updateBounds()

Do not call updateBounds when the bounds are either known or will be
overwritten. Add a resizeFill method that sets the size and contents
and does not call updateBounds.

Add LocalSelectivityVector interfaces that initialize the vector from
another vector or from size and value.

Fix SelectivityVector::== to only consider bits in between begin_ and
end_.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1672

Reviewed By: mbasmanova

Differential Revision: D36584889

Pulled By: oerling

fbshipit-source-id: 6557525296004860b2a91af4502bc9ac2cbc1d7b