Added script to alias include guards and updated readme with instructions
DuckDB includes headers from Thrift. The new parquet reader also needs to include the same Thrift headers. To avoid collisions during compilation we need to alias the include guard variables used in the header so that they don't conflict. The script finds all include guard variables that begin with '_THRIFT' and prepends a '_DUCKDB' to them.