Fix bug with != filter for integers and add corresponding tests (#1716)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1716

Added the `integerNotEqualFilter` test in `TableScanTest.cpp` for queries of the form `col != n` for integers `n`. Also added the `stringNotEqualFilter` test in `TableScanTest.cpp` for queries of the form `col != s` for strings `s`.

In `PlanBuilder.cpp`, fixed the `makeNotEqualFilter` function such that when it is being used on integers, no errors are returned. This also preserves the existing capability of the not-equal filter to check if strings and floating-point numbers are equal. This uses the `kBigintMultiRange` for integer filters and the normal `kMultiRange` for other types.

Also extended `makeLessThanOrEqualFilter` and similar functions in `PlanBuilder.cpp` to support `TINYINT` and `SMALLINT` types.

Reviewed By: gggrace14

Differential Revision: D36643745

fbshipit-source-id: 02591c8d2cb71ac151eb6fdaaa0df32d6879291f