Refactor dwrf::StreamIdentifier
dwrf::StreamIdentifier was used as a parameter of BufferedInput::
enqueue(), which causes circular dependency problems for future
refactoring that will move dwrf::common::BufferedInput to dwio::common.
This commit extracts a common interface and moves it to dwio::common::
StreamIdentifier. dwrf::DwrfStreamIdentifier, a subclass of dwio::
common::StreamIdentifier, is introduced for dwrf. The EncodingKey,
which was used to identify a DWRF encoding, is now a field of
DwrfStreamIdentifier.