json_extract_scalar() should always throw when path is invalid
Summary:
To follow the exact semantic in Presto, json_extract_scalar() should
always throw if the path provided is invalid. The previous code used to return
null in case the json was invalid.
Also refactoring the code to encapsulate more code into JsonExtractor and
remove it from the header, since it doesn't need to be part of the API.

Reviewed By: kagamiori

Differential Revision: D36394476

fbshipit-source-id: 3e789916b2f6ba5c32f3445a929860af67ba0005