Add fast path to Zip when all Arrays are flat and the same size (#1765)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1765

Zip pessimistically assumes that the Arrays for a given row
are different sizes.  This means it allocates nulls so it can fill
 in shorter arrays, and it dictionary encodes the elements
from the original arrays.

If all the input vectors are "flat" ArrayVectors (this is stricter
than what we need which is that they all have the same encoding, this is just a common, easy to check for case) and
for every row all the arrays within that row are the same length (no need to add nulls) we don't have to allocate nulls
or dictionary encode the elements.

This is a performance savings, since we have to do
significantly less work, and if the elements are themselves
flat, we can potentially take advantage of optimizations in
downstream UDFs for flat Vectors.

Differential Revision: D36914803

fbshipit-source-id: 4fb2d9707a071543d4f6638cce29128f04bd2b77