Fix FlatVector<StringView>::set (#1792)
Summary:
It is possible to have two FlatVectors in two threads share string buffers. In
this case, calling FlatVector::set from two threads at the same time creates a
race condition:

- [t1] FlatVector<StringView>::set(i, "longish string")
- [t2] FlatVector<StringView>::set(i, "longish string")
- [t1] FlatVector<StringView>::getBufferWithSpace(14): buffer size is 10, capacity 32 -> enough space, use that buffer
- [t2] FlatVector<StringView>::getBufferWithSpace(14): buffer size is 10, capacity 32 -> enough space, use that buffer
- [t1] append "longish string" to the buffer making buffer size 10 + 14 = 24
- [t1] try to append "longish string" to the buffer that now holds 24 bytes with 12 bytes available - FAILURE: size <= capacity_ (34 vs. 32)

To fix this, modify getBufferWithSpace to check that buffer has enough space and
is uniquely referenced. Otherwise, make a new buffer.

Before modifying a vector, the caller is expected to call
BaseVector::ensureWritable which checks if the vector and buffers are uniquely
referenced and if not allocates new vector and buffers. However, ensureWritable
doesn't check that string buffers are uniquely referenced. An alternative fix
would be to modify ensureWritable to check string buffers and make a copy if
these are not uniquely referenced.

The reason for making a change in FlatVector::set instead is because often
strings are written by copying StringViews and referencing the buffers, not by
writing the strings into the buffers. Hence, aggressively copying string
buffers in ensureWritable will perform costly copying unnecessarily.

```
template <>
void FlatVector<StringView>::copy(
    const BaseVector* source,
    const SelectivityVector& rows,
    const vector_size_t* toSourceRow) {
  ...

  auto leaf = source->wrappedVector()->asUnchecked<SimpleVector<StringView>>();

  if (pool_ == leaf->pool()) {
    // We copy referencing the storage of 'source'.
    copyValuesAndNulls(source, rows, toSourceRow);
    acquireSharedStringBuffers(source);
  } else {
    rows.applyToSelected([&](vector_size_t row) {
      auto sourceRow = toSourceRow ? toSourceRow[row] : row;
      if (source->isNullAt(sourceRow)) {
        setNull(row, true);
      } else {
        set(row, leaf->valueAt(source->wrappedIndex(sourceRow)));
      }
    });
  }
  ...
}
```

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1792

Reviewed By: spershin

Differential Revision: D37028776

Pulled By: mbasmanova

fbshipit-source-id: 8996ba0edcd1ee3c1865a0bd5e26ae334530763f