Fix signed integer overflow in range VectorHasher
A group by VectorHasher can overflow when adding reserve padding
around a range that is close to the end of its type's range
bounds. Bounds the VectorHasher range to the range of the column
type.

Moves the logic for padding value or id ranges to VectorHasher.cpp
from HashTable.cpp. Group by hash tables leave extra values at ends of
a range of values or ids to accommodate a certain amount of new keys
without rehashing. The amount by which the range can be extended
depends on the data types and how close the values are to the end of
the type's range, so the VectorHasher knows best.

Adds a test that verifies rounding stays within type bounds and
distinct normalized keys are produced for all combinations of nulls,
type min/max values of 8/16/32 bit types.