Fix seek of PagedInputStream over uncompressed data
PagedInputStream decompresses a whole compression entry and returns
windows into that. If the underlying data is not compressed, it
returns the windows it gets from its input. When seeking to a
position, if the target position falls in the current compression run,
we return a window into the uncompressed data starting at the desired
offset. This offset is anywhere between the start and end of the
current uncompressed buffer.

However, the compression run may be covered by multiple consecutive
windows into the input stream buffer.  If these windows of the are
directly returned to the caller, the seek to position will only work
for positions that fall in the current window, which may not include
the start of the compression run.

We add a check that we do not backup outside of the last window if
directly returning windows into the input's buffer. When seeking to
position, we distinguish the case of returning windows into the input
directly from the case of returning windows into the uncompressed
buffer.