Add NegatedValues filter classes for integers (#1782)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1782

This diff adds the `NegatedBigintValuesUsingBitmask` and `NegatedBigintValuesUsingHashTable` classes to `Filter.h` as wrappers for the existing `BigintValuesUsingBitmask` and `BigintValuesUsingHashTable` classes. These classes are meant to be used with filters in the form `NOT IN [list]` and `!= <value>`.

Implementations for the relevant `testInt64` and `testInt64Range` methods, as well as constructors methods have been added to `Filter.h` and `Filter.cpp`. Additionally, the `createNegatedBigintValues` function has been added to the `common` namespace, which takes a vector of values rejected by the filter and creates either a bitmask or hash-table NegatedValues filter as is appropriate for the given data set. These methods serve the same purpose as the `BigintValues` versions of these filters. The next thing to implement is the `mergeWith` methods, which will be created in a future diff.

Tests were also added to `FilterTest.cpp` for all of the newly-implemented classes, and benchmarks were added to `NegatedValuesFilterBenchmark.cpp`.

Differential Revision: D36974472

fbshipit-source-id: 42e315934a8673c997f90cf8871a871458961291