Add a SpeedTest utility binary (#1807)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1807

Add a SpeedTest utility binary to help evaluate and benchmark how
fast we can generate TPC-H datasets using the TPC-H Connector.

The first observation is that increasing the number of drivers doesn't increase throughput (due to the internal lock on dbgen). This binary will help evaluate the speed as we work on removing the limitation.

Reviewed By: mbasmanova

Differential Revision: D37020850

fbshipit-source-id: f6178d458dbad7be33c824fb655e2eae4cfc6780