Fix accounting of the total size of buffered data in PartitionedOutputBuffer (#1815)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1815

PartitionedOutputBuffer class is managing a set of output buffers for a given
task. There is one buffer per partition (corresponds 1:1 to a downstream
task).

Producers (PartitionedOutput operators) call PartitionedOutputBuffer::enqueue
() method to add data to buffers.

Consumers (downstream tasks) call PartitionedOutputBuffer::getData() and
PartitionedOutputBuffer::acknowledge() methods to fetch data and acknowledge
the receipt.

The data is stored in the buffers as a list of std::shared_ptr<SerializedPage>.
For repartitioning, each piece of data is stored in a single buffer and
corresponding shared_ptr is singly-referenced. For broadcast, each piece of
data is duplicated into all the buffers and corresponding shared_ptr have as
many references as there are buffers. As data is read and acknowledged from the
buffers, the reference count decreases. When the last "copy" is read, the
reference count goes to zero.

PartitionedOutputBuffer tracks total number of bytes of buffered data in a
totalSize_ member variable and blocks producers if that number exceeds a
threshold. As the data is read and removed from the buffers totalSize_
decreases. When totalSize_ goes below the threshold producers get unblocked.

A bug in calculating totalSize_ caused over-counting and indefinite blocking of
the producers resulting in a query getting stuck.

Calculation of totalSize_ relies on references count of the
std::shared_ptr<SerializedPage>. The number increases in enqueue() and decreases
in acknowledge(), but only if reference count goes to zero. Here is the
relevant code:

```
void PartitionedOutputBuffer::updateAfterAcknowledgeLocked(
    const std::vector<std::shared_ptr<SerializedPage>>& freed,
    std::vector<ContinuePromise>& promises) {
  uint64_t totalFreed = 0;
  for (const auto& free : freed) {
    if (free.unique()) {
      totalFreed += free->size();
    }
  }
  ...
  totalSize_ -= totalFreed;
  if (totalSize_ < continueSize_) {
    promises = std::move(promises_);
  }
```

PartitionedOutputBuffer::getData() used to create a second reference of the
std::shared_ptr<SerializedPage> stored in the buffer, pass it to the consumer
and continue holding it while freeing implicitly acknowledged data. This time
window was sufficint to allow the data to be received by the downstream task
and an ack being sent (the downstream task happens to run on the same machine).
While processing the ack, the reference count was not zero and totalSize_ was
not decreased effectively leaking so many bytes.

The fix is to avoid taking a second reference to the buffered data by changing the
getData() API to return a shallow copy of the folly::IOBufs pointing to the
data.

Differential Revision: D37163954

fbshipit-source-id: f9732af26a97d4d012d5b2da1f5dd62a75120a92