Refactor integer readers to share boilerplate
The template instantiation logic for direct and dictionary integer
readers is the same. The difference is at the top of read() where we
seek different streams and at the bottom where we call different
decoders on different visitors.

Adds a conversion from direct to dictionary visitor. This supports
switching between encodings in mid-stripe.

This allows directly inheriting the integer logic for Parquet and Alpha.