Add NegatedValues filter classes for integers (#1782)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1782

This diff adds the `NegatedBigintValuesUsingBitmask` and `NegatedBigintValuesUsingHashTable` classes to `Filter.h` as wrappers for the existing `BigintValuesUsingBitmask` and `BigintValuesUsingHashTable` classes. These classes are meant to be used with filters in the form `NOT IN [list]` and `!= <value>`.

Implementations for the relevant `testInt64` and `testInt64Range` methods, as well as constructors methods have been added to `Filter.h` and `Filter.cpp`. Additionally, the `createNegatedBigintValues` function has been added to the `common` namespace, which takes a vector of values rejected by the filter and creates either a bitmask or hash-table NegatedValues filter as is appropriate for the given data set. These methods serve the same purpose as the `BigintValues` versions of these filters. The next thing to implement is the `mergeWith` methods, which will be created in a future diff.

Tests were also added to `FilterTest.cpp` for all of the newly-implemented classes, and benchmarks were added to `NegatedValuesFilterBenchmark.cpp`.

Benchmark results with optimizations enabled: P509063862

The new filter performs better than the old filter and runs twice as fast on most cases, improving as the length of the list in the NOT IN predicate increases. It runs identically on the != cases, since neither the bitmask nor hash-table implementations will accept a set of rejected values of size 1 and therefore a multi-range filter is still being used there. I also ran the benchmarks again when using a modified bitmask for the != query (which does not reject a `values` vector of length 1) and it runs about 2-3 times faster than the current multiRange query, which is another possible source of improvement but would require modification to the `BigintValuesUsingBitmask` class, which is already in use elsewhere.

Reviewed By: Yuhta, gggrace14

Differential Revision: D36974472

fbshipit-source-id: 65e555b17a5511ab9379abbd0088e62ff2507e3c