Improve infer type for MAP (#1845)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1845

# Problem
`variant::inferType()` threw an exception if MAP could not infer types for key and value. Which is an inconsistent behavior to ARRAY type, that returns `ARRAY(UNKNOWN())` instead. I'd like to keep it consistent.

# Solution
Infer logic iterates for each element and get the type of the first not-null entry. I updated the logic to get key and value type separately, and if it can't it would use `UNKNOWN()` instead of throwing an exception.

I added some test cases to confirm the behavior. I hope this makes sense.

Reviewed By: abhash09

Differential Revision: D37313760

fbshipit-source-id: 9cbf1cab219da720c6ce84368fdc76d9758c3d7d