Use folly::Unit instead of bool for Driver continue future (#776)
Summary:
X-link: https://github.com/facebookexternal/presto_cpp/pull/776

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1843

A continue future can be used to suspend a driver and resume it once fulfilled,
right now a future<bool> is used, however, the bool value is
never read anywhere for any of the operators and never used.

Furthermore, it is a bit confusing as the user won't know if he shall set true or false,
the proper and official way of implementing a void future is by using future<unit>.

The diff also does the following:

1. use VeloxPromise for all continued promises (which are checked by
the Velox to be fulfilled before destructed or otherwise a warning printed).

2. Use the ContinueFuture alias for all continue futures, this changed some futures to semi-future.

3. avoid double initializing of futures in operators that do that.

Reviewed By: mbasmanova

Differential Revision: D37194084

fbshipit-source-id: b95e93d0d5e0611306193a9208f9e003fb5dbcea