Implement MergeWith for integer NegatedValues classes (#1834)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1834

This diff completes the implementation of `mergeWith` for the `NegatedBigintValuesUsingHashTable` and `NegatedBigintValuesUsingBitmask` classes that were introduced in diff D36974472 (https://github.com/facebookincubator/velox/commit/7522860bd455721153663706515ddd2712c5c41f). `MergeWith` is a function within these filter classes that takes a pointer to another filter and returns a filter that accepts values that pass both filters.

Test cases for the `MergeWith` function were also added to `FilterTest.cpp`. In this case, they were added by extending the `mergeWithBigint` and `mergeWithBigintMultiRange` unit tests, which test the capability of these filters to merge with other integer filters and checks the correctness. To accommodate this testing, an extra method `notIn` was also added to `ExprToSubfieldFilter.h` to easily create these filters.

Differential Revision: https://www.internalfb.com/diff/D37189376?entry_point=27

fbshipit-source-id: 65efc09fab34f18ad0a98f3abad6980dabed76ca