Add support for NegatedValues filters to dwio/dwrf/reader (#1835)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1835

Extended the dwrf reader to support the `NegatedBigintValuesUsingBitmask` and `NegatedBigintValuesUsingHashTable` classes of filters when reading data from a file. To do so, changes were made to `SelectiveIntegerDirectColumnReader.h`, `SelectiveIntegerDictionaryColumnReader.h`, and `SelectiveByteRleColumnReader.h` to support the new types of filters in the `processFilter` method. While the changes are not consequential along the current execution path, passing the information regarding the filter type to the `ReadHelper` template could be useful in future code changes.

The existing unit tests in `E2EFilterTest.cpp` were also modified to include NegatedValues test cases by modifying the `makeRangeFilter` method in `FilterGenerator.h`, which causes NegatedValues filters to be randomly generated sometimes.

Reviewed By: gggrace14

Differential Revision: D37133348

fbshipit-source-id: 0f31907cd095d8a5fa27cd2c7656e1c559a7821e