Add support for running tasks on the caller's thread (#1829)
Summary:
Allow the Task to be executed synchronously on the caller's thread instead of
folly::Executor specified in QueryCtx.

This mode of execution doesn't support arbitrary query plans yet. The following
restrictions apply:

- Query plan must translate to a single pipeline, e.g. no joins, no local
  exchanges, etc.
- If source operator requires splits, all splits must be added and
  no-more-splits flag must be set before starting the execution. This ensures
  that source operator won't block waiting for splits.
- No operator in the pipeline should block.
- The last operator (sink) must return data in Operator::getOutput, e.g.
  PartitionedOutputOperator or CallbackSink operators cannot be used.

To execute Task synchronously on the caller's thread, create the task, add all
splits, then call Task::next() in a loop to fetch data one batch at a time
until it returns a nullptr or throws an exception.

The new TaskTest::singleThreadedExecution shows how to use the new API.

We'll work on removing the restrictions above in follow-up PRs.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1829

Reviewed By: satya-valluri, oerling

Differential Revision: D37230311

Pulled By: mbasmanova

fbshipit-source-id: ae7d51e4d8ece61a03f133139bdfa883aeec9383