Make type checking in RowVector constructor debug-only (#1822)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1822

The constructor for RowVector checks that the Types of the children Vectors match the
children Types of the RowType passed in.

For very wide Rows with complex children this can become a significant performance
bottleneck.  Since it is not a critical check and mostly used for debugging, I'm limiting it to
a debug-only check so it doesn't impact production.

(Thanks Masha for the suggestion)

Reviewed By: mbasmanova

Differential Revision: D37196122

fbshipit-source-id: 07bf58c862310e8cf845976d3b29b8a237688079