Struct fields are captured as top level columns in lambdas (#1858)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1858

Lambdas currently capture all FieldAccesses that don't reference a parameter to the
lambda.  These are then treated as if they are accessing a column in the input data (e.g.
getting added to the distinctFields_ of the expression).

However, not all FieldAccesses are actually accessing input columns, they can also be
accessing fields of a struct.  If this is the case, the query will fail as it won't be able to find
the field in the input columns.

This change fixes this by checking to see if the FieldAccess is accessing an input column,
and only capturing it if so (e.g. if the struct is an input column that struct will be captured,
but accessing a field in that struct will not).

I also added a general purpose function to FieldAccessTypedExpr to help distinguish
between input columns and struct fields (I plan to do a similar check elsewhere and the
logic is slightly non-obvious so it's cleaner to just abstract it in one place).

Reviewed By: mbasmanova

Differential Revision: D37389547

fbshipit-source-id: 0fd6f2fd9403b38b69bc1e91f6ee82f36c6ea62a