Add additional info to MEM_CAP_EXCEEDED error
This patch adds a mechanism to the MemoryUsageTracker class that
allows its user to register a callback to add additional info to a
MEM_CAP_EXCEEDED error when encountered during a call to increase
reservation. Moreover, this mechanism is leveraged to add details
at the Task and Operator level in the memory tracker hierarchy.

Test plan:
Added tests to force this error and verify details added to the
error message.