Remove unnecessary and misleading RowVector::loadedChildAt method (#1878)
Summary:
RowVector::loadedChildAt was loading child vector only it is was a LazyVector,
but not if it was a LazyVector wrapped in one or more levels of dictionary.
This behavior is inconsistent with BaseVector::loadedVector and therefore is
misleading. Removed this method and replaced the callsites to use
RowVector::childAt + BaseVector::loadedVector or loadedVectorShared.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1878

Differential Revision: D37472097

Pulled By: mbasmanova

fbshipit-source-id: 72618d563305951aac2b95bb4fae5504450cff5d