Remove unnecessary and misleading RowVector::loadedChildAt method (#1878)
Summary:
RowVector::loadedChildAt was loading child vector only if it was a LazyVector,
but not if it was a LazyVector wrapped in one or more levels of dictionary.
This behavior is inconsistent with BaseVector::loadedVector and therefore is
misleading. Removed this method and replaced the callsites to use
RowVector::childAt + BaseVector::loadedVector or loadedVectorShared.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1878

Reviewed By: kagamiori

Differential Revision: D37472097

Pulled By: mbasmanova

fbshipit-source-id: f2eca3aefa40d4df80cb6827c7760ec303c86c73