Add supportsSingleThreadedExecution() function to Task
Summary:
The logic for deciding whether or not a Task supports single thread execution (i.e. calls to next()) is complicated and spread across the LocalPlanner and Driver classes.

This change adds a function supportsSingleThreadedExecution() to Task that makes it easy for users to check in a single place whether or not they can call next() on a Task, or whether they need to use the multi-threaded API.

Reviewed By: mbasmanova

Differential Revision: D37499915

fbshipit-source-id: ec3b29c0989cebd6b37f72e98e093ec7cf3fc911