Add supportsSingleThreadedExecution() function to Task (#1885)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1885

The logic for deciding whether or not a Task supports single thread execution (i.e. calls to next()) is complicated and spread across the LocalPlanner and Driver classes.

This change adds a function supportsSingleThreadedExecution() to Task that makes it easy for users to check in a single place whether or not they can call next() on a Task, or whether they need to use the multi-threaded API.

Reviewed By: mbasmanova

Differential Revision: D37499915

fbshipit-source-id: 6d6fdc88ca8f641601ad5298b698c5149fccdbc1