Print descriptive error message on benchmark jobs (#1887)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1887

Adding a more descriptive message to users in case main builds are broken. Benchmark runs for PRs compare with a baseline (based on main), so in case main is broken the benchmark regression test suite fails. Printing a more descriptive error message to alert users about what's happening, and asking them to rebase on a healthy main.

Differential Revision: D37508791

fbshipit-source-id: ebf7c4cb042194c632b2c5bcfa1962ad290149d4