Optimize KLL sketch memory usage for small groups
Summary:
When the sketch is created, currently we allocate the memory depending
on the K passed in.  When the actual number of elements being tracked is much
smaller than K (often happening with group by aggregation), the unused memory is
wasted.  We optimize for this case by allowing the actual retained elements to
grow from empty instead of an initial size of K.

Differential Revision: D37182402

fbshipit-source-id: 025fe5e87b54c11186122f9082fc7915928ede90