Fix rowcount generation for lineitem
Summary:
Lineitem is generated based on rows counts from the Orders table, so
the offsets need to advance based on the requested rows, rather than on number
of returned rows in the vectors. This is more general and works for other
tables as well, since they only case when less rows are returned than requested
is when the connector exhausted the generated rows.

Differential Revision: D37584637

fbshipit-source-id: ea68b4dd81ac96314d5389165ee923bb8dce45b7