Fix buffer reuse bug on Upper/Lower function
Summary:
Another bug caught by Fuzzer: UpperLower vectorized function was
moving away the input buffer and losing all references to it, causing the
Vector that holds the string buffer to run out of scope and be destroyed before
the function execution started, causing the function to trigger a
use-after-free ASAN crash. The issue is only reported in ASAN mode.

Differential Revision: D37674304

fbshipit-source-id: efb450aa20b8a2450eaa749c7ee057c9c8dbeb41