Lazily throw exception in simple functions initialize() (#1957)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1957

Lazily throwing exceptions in simple functions initialize().

If the initialize() method provided by simple functions throw, we can't let the
exception bubble up immediately; rather, we need to capture the exception and
only throw it if there are any active rows.
.
Consider the function "f(p1, c1)" where c1 is a constant that makes f()
throw on initialize(). If we throw immediately on initialize() and p1 is
composed only of nulls, the expected behavior would be to optimize this
function out and return null, not to throw. This subtle semantic issue
has been reported by Fuzzer runs.

Reviewed By: kevinwilfong

Differential Revision: D37666610

fbshipit-source-id: eedcdcae2bb394342752021bce514657dc9b894b