Lazily throw exception in simple functions initialize()
Summary:
Lazily throw exception in simple functions initialize(). The issue is
that initialize() is called during expression compilation, so

If the initialize() method provided by simple functions throw, we can't let the
exception bubble up immediately; rather we need to capture the exception and
only throw it if there are any active rows.
.
Consider the function "f(p1, c1)" where c1 is a constant that makes f()
throw on initialize(). If we throw imediately on initialize() and p1 is
composed only of nulls, the expected behavior would be to optimize this
function out and return null, not to throw. This subtle semantic issue
has been reported by Fuzzer runs.

Differential Revision: D37666610

fbshipit-source-id: 257b2ccd10a6e45144474bf334a4284302dc5d02