Create micro-benchmarks for NegatedBytesValues filters (#1973)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1973

This diff adds a benchmark file to Velox for the new `NegatedBytesValues` class, which handles `NOT IN [list]` queries on strings. The benchmarks compare the old `MultiRange` filters, which previously were used for these types of queries, to the new `NegatedBytesValues` filters. To do so, the filters are run on randomly generated strings of varying length (1, 2, 5, 10, 100) and `NOT IN` lists of varying length (1, 5, 10, 100, 1000, 10000). The `MultiRange` filters generate `n` ranges on a `NOT IN` list of length `n`, and scan through all such ranges taking linear time. The `NegatedBytesValues` use a `folly::F14FastSet` (hash set) to get what should be constant time.

The results can be found in paste P512608110. From this, we can see that the `NegatedBytesValues` filters outperform the `MultiRange` filters on every input, running about twice as fast on strings of length 1 and `NOT IN` lists of medium size. As the `NOT IN` lists got longer, the `NegatedBytesValues` filters performed better relative to the `MultiRange` filters, working over 1500x faster in the most extreme conditions of length-10000 lists. However, as the strings got longer, the `NegatedBytesValues` filters offered slightly less of an advantage -- likely because the overhead of comparing strings gets more expensive as the strings get longer, and both types of filters must do these comparisons.

Reviewed By: gggrace14

Differential Revision: D37473861

fbshipit-source-id: 8b152d9097c03c32dd291ab1d65e3b4552b86512