Create NegatedBigintRange filter class (#1966)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1966

This diff adds the `NegatedBigintRange` class of filters, which represents queries in the form `NOT BETWEEN low AND high` for big integer bounds. The necessary methods for a string filter have been added to `Filter.h` and `Filter.cpp`, which builds a wrapper class around the existing `BigintRange` filter for BETWEEN queries. This diff also adds tests for the new class and its member functions within `FilterTest.cpp`, and modifies `ExprToSubfieldFilter.h` to create methods capable of generating these new filters for testing.

Reviewed By: Yuhta

Differential Revision: D37496231

fbshipit-source-id: baab05e7cefe02426a5fc3a1a637f55963f39e6a