Add Arrow Parquet Writer (#1991)
Summary:
Adds a thin wrapper around Arrow Parquet writer. Converts Velox
Vectors to Arrow using the Velox-Arrow bridge and then writes these
using the Arrow writer.

Adds the initial base class for the Parquet E2EFilterTest and a test
for checking that the writer produces the right magic numbers. As
parts of the reader are added we will check that the metadata and
columns are correctly read back.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1991

Differential Revision: D37786698

Pulled By: Yuhta

fbshipit-source-id: 5de12fa80c1e1585e6e15c8cd02ef59fce34cd51