Remove SelectiveColumnReader inheritance from ColumnReader
SelectiveColumnReader will be moved to dwio::common but ColumnReader
will remain untouched, based on the assumption ColumnReader and its
descendent classes will retire soon. This commit removes the inheritance
of SelectiveColumnReader to ColumnReader. To do this, DwrfRowReaderShared
would contain a SelectiveColumnReader and a ColumnReader as its members,
but only one of them is not null. TestColumnReader.cpp was also modified
this way to make the tests work. In addition, the contents in
TestColumnReader.cpp were reordered to form 3 sections: 1) test class
definitions, 2) tests, and 3) test instantiations.