Fix bug with NegatedBytesValues::MergeWith empty string handling
Summary: Previously, the `NegatedBytesValue::MergeWith` function would incorrectly attempt to remove the bounds reported by an unbounded `BytesRange` from the set of values rejected from the merged filter, when it should stay in the rejected set. A filter that is unbounded below should not have any lower bound, and this diff adds code to ignore the reported lower bound from such a `BytesRange` filter.

Reviewed By: gggrace14

Differential Revision: D37837207

fbshipit-source-id: 4115108c80f39bfb3a10725fd8a90aed42751c32